% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Distribution}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
barplot and table of the relative frequency
}
\description{
Sampling distribution showing relative frequency of the binomial samples with the given number of success
}
\examples{
mybin(iter=100, n=10, p=0.5)
}
