% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Boot function for sampling distributions}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iteration}

\item{x}{sample vector}

\item{fun}{function to apply}

\item{alpha}{error probability}

\item{cx}{1.5}
}
\value{
histogram of the functions
}
\description{
Boot function for sampling distributions
}
\examples{
myboot2(10000,x=rnorm(25,mean=25,sd=10),fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)
}
