% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Central Limit Thereom apply to binomal distribution}
\usage{
mycltb(n, iter, p)
}
\arguments{
\item{n}{number of trials}

\item{iter}{number of iterations}

\item{p}{probability of success}
}
\value{
histogram of binomial sampling distributions
}
\description{
Central Limit Thereom apply to binomal distribution
}
