% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max liklihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function parameter}

\item{x}{sample}

\item{param}{parameter}
}
\value{
return i, parameter, yi, and slope
}
\description{
Max liklihood function
}
\examples{
mymaxlik(lfun = function(x, param) log(dbinom(x, prob = param, size = 10)), x = c(3, 3, 4, 3, 4, 5, 5, 4), param = seq(0,1, length=1000, main="Binomial, n = 20"))
}
