% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve}
\usage{
myncurve(q, mu = 0, sigma = 1)
}
\arguments{
\item{q}{quantile}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
plot and probability of density curved
}
\description{
Normal Curve
}
\examples{
myncurve(2)
}
