% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculates the number of tickets to be sold when given the number of seats, probability of attendance, and probability of overbooking}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{probability of overbooking}

\item{p}{probability of attendence}
}
\value{
two plots of discrete distribution and continous distibution
}
\description{
Calculates the number of tickets to be sold when given the number of seats, probability of attendance, and probability of overbooking
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.05)
}
